payroll/backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
python-dotenv==1.0.0
PyMySQL==1.1.0
cryptography==41.0.7
Werkzeug==3.0.1
PyPDF2==3.0.1
reportlab==4.0.9
